requirements.txt
fastapi
uvicorn
databases[sqlite]
sqlalchemy
pydantic
Werkzeug
Six
python-dotenv
PyJWT